refactor(Display): 패키지 분리
