+++ src/main/java/view/display/CardShapeDisplay.java
@@ -1,4 +1,4 @@
-package view;
+package view.display;
 
 import card.Shape;
